+++ app/models/transaction_model.py
@@ -1,5 +1,7 @@
 from sqlalchemy.orm import Mapped, mapped_column, relationship
-from sqlalchemy import String, Integer, ForeignKey, DECIMAL, Numeric, func
+from sqlalchemy import String, Integer, ForeignKey, Numeric, func
+
+from decimal import Decimal
 
 from datetime import datetime
 
@@ -13,6 +15,6 @@ class TransactionModel(Base):
     id: Mapped[int] = mapped_column(primary_key=True)
     from_account_id: Mapped[int] = mapped_column(ForeignKey("accounts.id"))
     to_account_id: Mapped[int] = mapped_column(ForeignKey("accounts.id"))
-    amount:Mapped[float] = mapped_column(Numeric(20, 2))   
+    amount:Mapped[Numeric] = mapped_column(Numeric(20, 2))   
     
     created_at:Mapped[datetime] = mapped_column(server_default=func.now())
